% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_game_of_life.R
\name{calc_game_of_life}
\alias{calc_game_of_life}
\title{Calculate Alive or Dead Cells}
\usage{
calc_game_of_life(n_size, n_generations)
}
\arguments{
\item{n_size}{one side of the grid (dimensions n_size * n_size).}

\item{n_generations}{number of generations to simulate.}
}
\value{
List of cell values over generations.
}
\description{
Function calculates which cells survive, are born or die over generations
}
\examples{
gol_100_10 <- calc_game_of_life(
  n_size=100,
  n_generations=10
)

}
