% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_game_of_life.R
\name{draw_game_of_life}
\alias{draw_game_of_life}
\title{Plot the different generations}
\usage{
draw_game_of_life(n_size, n_generations, board_list)
}
\arguments{
\item{n_size}{one side of the grid (dimensions n_size * n_size).}

\item{n_generations}{number of generations of cells.}

\item{board_list}{of values (alive or dead cells) obtained from calc_game_of_life.}
}
\value{
Number of plots equal to the number of generations.
}
\description{
Function that displays a number of plots equal to the generations.
It requires the output generated from calc_game_of_life.
}
\examples{
draw_game_of_life(
  n_size=100,
  n_generations=10,
  board_list= gameoflife:::calc_game_of_life(
    n_size=100,
    n_generations=10
    )
)

}
