% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\docType{package}
\name{gameoflife}
\alias{gameoflife}
\title{Conway Game of Life}
\description{
This package contains two functions; one calculating which cells survive or die
from one generation to the next. The other one is used for drawing them in ggplot2.
}
\author{
Tomo Umer \email{tomo.umer@gmail.com}
}
