% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tick.R
\name{tick}
\alias{tick}
\title{Game of life tick}
\usage{
tick(cells)
}
\arguments{
\item{cells}{Current generation of cells}
}
\value{
A list of cells representing the next generation
}
\description{
Generates the next generation of cells.
}
\examples{

tick(list()) # list()

}
